Updated initial home HTML page
Oh my lord this is incredibly tedious and difficult
to do. The exercise calls for adding all the HTML content
in JavaScript by creating and appending content,
so I'm effectively doing everything in a JS and CSS file.
It's very confusing having to only reference the DOM
in the browser for the CSS and having to constantly read
my JS when I want to change everything. Very messy!